refactor(skill-tree): move concentric circles rendering to a new file and update imports
