+++ src/modules/skill-tree/component/ConcentricCircles.tsx
@@ -1,4 +1,4 @@
-import type { SkillTreeOptions } from "../../skill-tree.type";
+import type { SkillTreeOptions } from "../skill-tree.type";
 
 type ConcentricCirclesProps = {
   rings: number[];
